[listener] add instagram split1
